+++ core/src/main/kotlin/net/corda/core/internal/datavending/FetchDataFlow.kt
@@ -1,4 +1,4 @@
-package net.corda.flows
+package net.corda.core.internal.datavending
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.contracts.NamedByHash
@@ -6,11 +6,11 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.flows.FlowException
 import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.Party
+import net.corda.core.internal.datavending.FetchDataFlow.DownloadedVsRequestedDataMismatch
+import net.corda.core.internal.datavending.FetchDataFlow.HashNotFound
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.utilities.UntrustworthyData
 import net.corda.core.utilities.unwrap
-import net.corda.flows.FetchDataFlow.DownloadedVsRequestedDataMismatch
-import net.corda.flows.FetchDataFlow.HashNotFound
 import java.util.*
 
 /**
@@ -31,7 +31,8 @@ import java.util.*
  */
 abstract class FetchDataFlow<T : NamedByHash, in W : Any>(
         protected val requests: Set<SecureHash>,
-        protected val otherSide: Party) : FlowLogic<FetchDataFlow.Result<T>>() {
+        protected val otherSide: Party,
+        private val fetchDataType: DataType) : FlowLogic<FetchDataFlow.Result<T>>() {
 
     @CordaSerializable
     class DownloadedVsRequestedDataMismatch(val requested: SecureHash, val got: SecureHash) : IllegalArgumentException()
@@ -42,10 +43,18 @@ abstract class FetchDataFlow<T : NamedByHash, in W : Any>(
     class HashNotFound(val requested: SecureHash) : FlowException()
 
     @CordaSerializable
-    data class Request(val hashes: List<SecureHash>)
+    data class Result<out T : NamedByHash>(val fromDisk: List<T>, val downloaded: List<T>)
 
     @CordaSerializable
-    data class Result<out T : NamedByHash>(val fromDisk: List<T>, val downloaded: List<T>)
+    interface Request
+
+    data class DataRequest(val hashes: List<SecureHash>, val dataType: DataType) : Request
+    object EndRequest : Request
+
+    @CordaSerializable
+    enum class DataType {
+        TRANSACTION, ATTACHMENT
+    }
 
     @Suspendable
     @Throws(HashNotFound::class)
@@ -53,18 +62,19 @@ abstract class FetchDataFlow<T : NamedByHash, in W : Any>(
         // Load the items we have from disk and figure out which we're missing.
         val (fromDisk, toFetch) = loadWhatWeHave()
 
-        return if (toFetch.isEmpty()) {
+        val result = if (toFetch.isEmpty()) {
             Result(fromDisk, emptyList())
         } else {
             logger.trace("Requesting ${toFetch.size} dependency(s) for verification")
 
             // TODO: Support "large message" response streaming so response sizes are not limited by RAM.
-            val maybeItems = sendAndReceive<ArrayList<W>>(otherSide, Request(toFetch))
+            val maybeItems = sendAndReceive<ArrayList<W>>(otherSide, DataRequest(toFetch, fetchDataType))
             // Check for a buggy/malicious peer answering with something that we didn't ask for.
             val downloaded = validateFetchResponse(maybeItems, toFetch)
             maybeWriteToDisk(downloaded)
             Result(fromDisk, downloaded)
         }
+        return result
     }
 
     protected open fun maybeWriteToDisk(downloaded: List<T>) {
@@ -89,17 +99,14 @@ abstract class FetchDataFlow<T : NamedByHash, in W : Any>(
     @Suppress("UNCHECKED_CAST")
     protected open fun convert(wire: W): T = wire as T
 
-    private fun validateFetchResponse(maybeItems: UntrustworthyData<ArrayList<W>>,
-                                      requests: List<SecureHash>): List<T> {
+    private fun validateFetchResponse(maybeItems: UntrustworthyData<ArrayList<W>>, requests: List<SecureHash>): List<T> {
         return maybeItems.unwrap { response ->
-            if (response.size != requests.size)
-                throw DownloadedVsRequestedSizeMismatch(requests.size, response.size)
+            if (response.size != requests.size) throw DownloadedVsRequestedSizeMismatch(requests.size, response.size)
             val answers = response.map { convert(it) }
             // Check transactions actually hash to what we requested, if this fails the remote node
             // is a malicious flow violator or buggy.
             for ((index, item) in answers.withIndex()) {
-                if (item.id != requests[index])
-                    throw DownloadedVsRequestedDataMismatch(requests[index], item.id)
+                if (item.id != requests[index]) throw DownloadedVsRequestedDataMismatch(requests[index], item.id)
             }
             answers
         }