+++ core/src/main/kotlin/net/corda/flows/BroadcastTransactionFlow.kt
@@ -28,7 +28,8 @@ class BroadcastTransactionFlow(val notarisedTransaction: SignedTransaction,
         val msg = NotifyTxRequest(notarisedTransaction)
         participants.filter { it != serviceHub.myInfo.legalIdentity }.forEach { participant ->
             // This pops out the other side in NotifyTransactionHandler
-            send(participant, msg)
+            // SendTransactionFlow allows otherParty to access our data to resolve the transaction.
+            subFlow(SendTransactionFlow(participant, msg))
         }
     }
 }