+++ core/src/main/kotlin/net/corda/flows/ResolveTransactionsFlow.kt
@@ -1,22 +1,25 @@
 package net.corda.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.checkedAdd
 import net.corda.core.crypto.SecureHash
 import net.corda.core.flows.FlowLogic
 import net.corda.core.getOrThrow
 import net.corda.core.identity.Party
+import net.corda.core.internal.datavending.FetchAttachmentsFlow
+import net.corda.core.internal.datavending.FetchDataFlow
+import net.corda.core.internal.datavending.FetchTransactionsFlow
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction
 import java.util.*
-
-// TODO: This code is currently unit tested by TwoPartyTradeFlowTests, it should have its own tests.
+import kotlin.collections.LinkedHashSet
 
 // TODO: It may be a clearer API if we make the primary c'tor private here, and only allow a single tx to be "resolved".
 
 /**
+ * The [ResolveTransactionsFlow] corresponds to the [SendTransactionFlow].
+ *
  * This flow is used to verify the validity of a transaction by recursively checking the validity of all the
  * dependencies. Once a transaction is checked it's inserted into local storage so it can be relayed and won't be
  * checked again.
@@ -34,19 +37,15 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
 
     companion object {
         private fun dependencyIDs(wtx: WireTransaction) = wtx.inputs.map { it.txhash }.toSet()
-
         /**
          * Topologically sorts the given transactions such that dependencies are listed before dependers. */
         @JvmStatic
         fun topologicalSort(transactions: Collection<SignedTransaction>): List<SignedTransaction> {
             // Construct txhash -> dependent-txs map
-            val forwardGraph = HashMap<SecureHash, HashSet<SignedTransaction>>()
-            transactions.forEach { stx ->
-                stx.tx.inputs.forEach { (txhash) ->
+            val forwardGraph = transactions.flatMap { stx -> stx.tx.inputs.map { it.txhash to stx } }
+                    .groupBy { it.first }
                     // Note that we use a LinkedHashSet here to make the traversal deterministic (as long as the input list is)
-                    forwardGraph.getOrPut(txhash) { LinkedHashSet() }.add(stx)
-                }
-            }
+                    .mapValues { LinkedHashSet(it.value.map { it.second }) }
 
             val visited = HashSet<SecureHash>(transactions.size)
             val result = ArrayList<SignedTransaction>(transactions.size)
@@ -58,14 +57,11 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
                     result.add(transaction)
                 }
             }
-
             transactions.forEach(::visit)
-
             result.reverse()
             require(result.size == transactions.size)
             return result
         }
-
     }
 
     @CordaSerializable
@@ -96,43 +92,33 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
     @Suspendable
     @Throws(FetchDataFlow.HashNotFound::class)
     override fun call(): List<LedgerTransaction> {
-        val newTxns: Iterable<SignedTransaction> = topologicalSort(downloadDependencies(txHashes))
-
-        // For each transaction, verify it and insert it into the database. As we are iterating over them in a
-        // depth-first order, we should not encounter any verification failures due to missing data. If we fail
-        // half way through, it's no big deal, although it might result in us attempting to re-download data
-        // redundantly next time we attempt verification.
-        val result = ArrayList<LedgerTransaction>()
-
-        for (stx in newTxns) {
-            // Resolve to a LedgerTransaction and then run all contracts.
-            val ltx = stx.toLedgerTransaction(serviceHub)
-            // Block on each verification request.
-            // TODO We could recover some parallelism from the dependency graph.
-            serviceHub.transactionVerifierService.verify(ltx).getOrThrow()
-            serviceHub.recordTransactions(stx)
-            result += ltx
-        }
-
-        // If this flow is resolving a specific transaction, make sure we have its attachments and then verify
-        // it as well, but don't insert to the database. Note that when we were given a SignedTransaction (stx != null)
-        // we *could* insert, because successful verification implies we have everything we need here, and it might
-        // be a clearer API if we do that. But for consistency with the other c'tor we currently do not.
-        //
-        // If 'stx' is set, then 'wtx' is the contents (from the c'tor).
         val wtx = stx?.verifySignatures() ?: wtx
-        wtx?.let {
-            fetchMissingAttachments(listOf(it))
-            val ltx = it.toLedgerTransaction(serviceHub)
-            ltx.verify()
-            result += ltx
-        }
 
-        return result
+        // Start fetching data.
+        val newTxns = downloadDependencies(txHashes)
+        fetchMissingAttachments(newTxns.flatMap { it.tx.attachments } + (wtx?.attachments ?: emptyList()))
+        send(otherSide, FetchDataFlow.EndRequest)
+        // Finish fetching data.
+
+        val result = topologicalSort(newTxns).map {
+            // For each transaction, verify it and insert it into the database. As we are iterating over them in a
+            // depth-first order, we should not encounter any verification failures due to missing data. If we fail
+            // half way through, it's no big deal, although it might result in us attempting to re-download data
+            // redundantly next time we attempt verification.
+            val ltx = it.toLedgerTransaction(serviceHub)
+            serviceHub.transactionVerifierService.verify(ltx).getOrThrow()
+            serviceHub.recordTransactions(it)
+            ltx
+        }
+        return if (wtx == null) {
+            result
+        } else {
+            result + wtx.toLedgerTransaction(serviceHub).apply { verify() }
+        }
     }
 
     @Suspendable
-    private fun downloadDependencies(depsToCheck: Set<SecureHash>): Collection<SignedTransaction> {
+    private fun downloadDependencies(depsToCheck: Set<SecureHash>): List<SignedTransaction> {
         // Maintain a work queue of all hashes to load/download, initialised with our starting set. Then do a breadth
         // first traversal across the dependency graph.
         //
@@ -153,36 +139,32 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
         nextRequests.addAll(depsToCheck)
         val resultQ = LinkedHashMap<SecureHash, SignedTransaction>()
 
-        val limit = transactionCountLimit
-        check(limit > 0) { "$limit is not a valid count limit" }
-        var limitCounter = 0
+        var limitCounter = transactionCountLimit
+        check(limitCounter > 0) { "$limitCounter is not a valid count limit" }
+
         while (nextRequests.isNotEmpty()) {
+            if (limitCounter < 0) throw ExcessivelyLargeTransactionGraph()
+            limitCounter -= nextRequests.size
+
             // Don't re-download the same tx when we haven't verified it yet but it's referenced multiple times in the
             // graph we're traversing.
             val notAlreadyFetched = nextRequests.filterNot { it in resultQ }.toSet()
-            nextRequests.clear()
-
-            if (notAlreadyFetched.isEmpty())   // Done early.
-                break
+            if (notAlreadyFetched.isEmpty())
+                break // Done early.
 
             // Request the standalone transaction data (which may refer to things we don't yet have).
-            val downloads: List<SignedTransaction> = subFlow(FetchTransactionsFlow(notAlreadyFetched, otherSide)).downloaded
-
-            fetchMissingAttachments(downloads.map { it.tx })
+            val downloads = subFlow(FetchTransactionsFlow(notAlreadyFetched, otherSide)).downloaded
 
             for (stx in downloads)
                 check(resultQ.putIfAbsent(stx.id, stx) == null)   // Assert checks the filter at the start.
 
             // Add all input states to the work queue.
             val inputHashes = downloads.flatMap { it.tx.inputs }.map { it.txhash }
+
+            nextRequests.clear()
             nextRequests.addAll(inputHashes)
-
-            limitCounter = limitCounter checkedAdd nextRequests.size
-            if (limitCounter > limit)
-                throw ExcessivelyLargeTransactionGraph()
         }
-
-        return resultQ.values
+        return resultQ.values.toList()
     }
 
     /**
@@ -190,11 +172,9 @@ class ResolveTransactionsFlow(private val txHashes: Set<SecureHash>,
      * first in the returned list and thus doesn't have any unverified dependencies.
      */
     @Suspendable
-    private fun fetchMissingAttachments(downloads: List<WireTransaction>) {
+    private fun fetchMissingAttachments(downloads: List<SecureHash>) {
         // TODO: This could be done in parallel with other fetches for extra speed.
-        val missingAttachments = downloads.flatMap { wtx ->
-            wtx.attachments.filter { serviceHub.attachments.openAttachment(it) == null }
-        }
+        val missingAttachments = downloads.filter { serviceHub.attachments.openAttachment(it) == null }
         if (missingAttachments.isNotEmpty())
             subFlow(FetchAttachmentsFlow(missingAttachments.toSet(), otherSide))
     }